refactor(orchestration): split scheduler.rs into focused sub-modules
Split 4627-line scheduler.rs into four focused modules with no behavior change:
- scheduler/mod.rs: public types, DagScheduler struct, constructors, accessors
- scheduler/tick.rs: core tick loop, event processing, spawn record-keeping
- scheduler/planner.rs: topology analysis, level barrier, graph completion checks
- scheduler/router.rs: task prompt construction, dependency context injection
- scheduler/persistence.rs: task injection, predicate outcome recording

All methods use direct impl DagScheduler blocks (no traits) so callers require
no import changes. All 337 zeph-orchestration unit tests continue to pass.

Also add deferred TODO(arch-revised-2026-04-26) markers in:
- zeph-core/src/agent/mod.rs: agent split blocked on task_supervisor adoption
- zeph-llm/src/any.rs: AnyProvider deprecation deferred to m49+
- zeph-tui/src/app/mod.rs: TUI Reducer/Action decomposition deferred to m49+
